requirements.txt
protobuf==4.25.2
google-generativeai==0.3.2
streamlit==1.11.0
python-dotenv==0.19.4
google==3.0.0
